pyproject.toml
[project]
name = "intelligent-search-agent"
version = "0.1.0"
description = "Assistant-first RAG API with asset, document, image, and meeting search foundations."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "uvicorn[standard]>=0.38.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.12.0",
    "pydantic-ai-slim[openai]>=1.34.0",
    "psycopg2-binary>=2.9.11",
    "httpx>=0.28.0",
    "python-dotenv>=1.2.0",
    "python-multipart>=0.0.20",
    "pymupdf>=1.24.0",
    "pytesseract>=0.3.13",
    "pillow>=10.0.0",
    "tzdata; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]
ingestion = [
    "pandas>=2.0.0",
]

[tool.setuptools.packages.find]
include = ["intelligent_search_agent*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
